Split error types out of form error handling code so they can be re-used for query errors
