+++ src/api/form-error-handling.ts
@@ -1,74 +1,7 @@
-import { ApolloError } from '@apollo/client';
 import { FORM_ERROR, FormApi, setIn, SubmissionErrors } from 'final-form';
 import { identity, mapValues } from 'lodash';
-import { assert } from 'ts-essentials';
 import { Promisable } from 'type-fest';
-
-/**
- * This is a mapping of error codes to their error object.
- * This mapping should be expanded upon as we add and handle
- * new server errors.
- */
-interface ErrorMap {
-  Validation: ValidationError;
-  NotFound: InputError;
-  TokenInvalid: ErrorInfo;
-  TokenExpired: ErrorInfo;
-  Input: InputError;
-  Duplicate: DuplicateError;
-
-  /**
-   * This is a special one that allows a default handler for any
-   * un-handled error codes.
-   */
-  Client: ErrorInfo;
-  /**
-   * This is a special one that handles server errors. i.e. 500
-   */
-  Server: ErrorInfo;
-
-  /**
-   * This is a special one that allows a default handler for any
-   * un-handled error codes.
-   */
-  Default: ErrorInfo;
-}
-
-type Code = keyof ErrorMap;
-
-/**
- * The basic error shape
- */
-interface ErrorInfo {
-  message: string;
-}
-
-/**
- * Validation errors also give an errors object which contains
- * the fields errors.
- */
-export interface ValidationError extends ErrorInfo {
-  /**
-   * All of the invalid fields and their errors.
-   * Nested fields are flattened to `a.b` keys which is great for final-form.
-   *
-   * @example
-   * {
-   *   'user.email': {
-   *     isEmail: 'email is not an email',
-   *     ...
-   *   },
-   *   ...
-   * }
-   */
-  errors: Record<string, Record<string, string>>;
-}
-
-export interface InputError extends ErrorInfo {
-  field?: string;
-}
-
-export type DuplicateError = Required<InputError>;
+import { ErrorMap, getErrorInfo, ValidationError } from './error.types';
 
 /**
  * A mapping where the key is the error code from the server
@@ -187,27 +120,6 @@ export const handleFormError = async <T, P>(
       (() => undefined) as NextHandler<any>
     );
   return handler(error);
-};
-
-const getErrorInfo = (e: unknown) => {
-  if (!(e instanceof ApolloError) || e.graphQLErrors.length === 0) {
-    // This is really to make TS happy. We should always have an ApolloError here.
-    assert(e instanceof Error);
-    return {
-      message: e.message,
-      codes: ['Default'] as Code[],
-    };
-  }
-
-  // For mutations we will assume they will only have one error
-  // since they should only be doing one operation.
-  const ext = e.graphQLErrors[0].extensions ?? {};
-  const codes: Code[] = [...(ext.codes ?? [ext.code]), 'Default'];
-  return {
-    message: e.message,
-    ...ext,
-    codes,
-  };
 };
 
 /**
